Add visual icons in problem list.
Add icons for each judge that is used in the dashboard. This icons are shown in the problem list beside the link to access to each problem.
Also, the file Data.js was modified to add the path to the icons that are used.

Icons added are alocated in the path:
  - app/assets/img